Move more classes out of Resolver.cs
